requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<5.0


